+++ agents/processor/src/processor.rs
@@ -423,7 +423,8 @@ impl NomadAgent for Processor {
             // if we have a bucket, add a task to push to it
             if let Some(config) = &self.config {
                 info!(bucket = %config.bucket, "Starting S3 push tasks");
-                tasks.push(Pusher::new(self.core.home.name(), &config.bucket, db.clone()).spawn())
+                let pusher = Pusher::new(self.core.home.name(), &config.bucket, db.clone()).await;
+                tasks.push(pusher.spawn())
             }
 
             // find the first task to shut down. Then cancel all others