+++ nomad-core/src/signer.rs
@@ -13,10 +13,6 @@ use ethers::{
     signers::{AwsSignerError, LocalWallet, Signer},
 };
 use nomad_xyz_configuration::agent::SignerConf;
-use once_cell::sync::Lazy;
-use rusoto_kms::KmsClient;
-
-static KMS_CLIENT: Lazy<KmsClient> = Lazy::new(|| KmsClient::new(Default::default()));
 
 /// Error types for Signers
 #[derive(Debug, thiserror::Error)]
@@ -62,7 +58,8 @@ impl Signers {
         match conf {
             SignerConf::HexKey(key) => Ok(Self::Local(key.as_ref().parse()?)),
             SignerConf::Aws { id } => {
-                let signer = AwsSigner::new(&KMS_CLIENT, id, 0).await?;
+                let kms_client = crate::aws::get_kms_client().await;
+                let signer = AwsSigner::new(kms_client, id, 0).await?;
                 Ok(Self::Aws(signer))
             }
             SignerConf::Node => bail!("Node signer"),