+++ nomad-core/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ### Unreleased
 
+- refactor: Add IRSA credentials to client instantiation
 - implement `Encode` and `Decode` for `bool`
 - adds a changelog